refactor(db): make the db logic a separate package
